[expo-router][docs] separate expo-rotuer reference into multiple subsections
