+++ docs/pages/versions/unversioned/sdk/router/color.mdx
@@ -1,0 +1,35 @@
+---
+title: Color
+description: Platform color utilities for Expo Router.
+sourceCodeUrl: 'https://github.com/expo/expo/tree/main/packages/expo-router'
+packageName: 'expo-router'
+platforms: ['android', 'ios']
+---
+
+import APISection from '~/components/plugins/APISection';
+
+> See the [Expo Router](./index) reference for installation and configuration.
+
+## Usage
+
+```tsx
+import { Color } from 'expo-router';
+import { Text, View, useColorScheme } from 'react-native';
+
+export default function MyComponent() {
+  useColorScheme();
+  return (
+    <View style={{ flex: 1, backgroundColor: Color.android.dynamic.primary }}>
+      <Text style={{ color: Color.ios.label }}>Hello</Text>
+    </View>
+  );
+}
+```
+
+## API
+
+```js
+import { Color } from 'expo-router';
+```
+
+<APISection packageName="expo-router/color" apiName="Color" />
