+++ docs/pages/versions/unversioned/sdk/router/link.mdx
@@ -1,0 +1,41 @@
+---
+title: Link
+description: Link, Redirect, preview, and zoom transition components for Expo Router.
+sourceCodeUrl: 'https://github.com/expo/expo/tree/main/packages/expo-router'
+packageName: 'expo-router'
+platforms: ['android', 'ios', 'tvos', 'web', 'expo-go']
+---
+
+import { BookOpen02Icon } from '@expo/styleguide-icons/outline/BookOpen02Icon';
+
+import APISection from '~/components/plugins/APISection';
+import { BoxLink } from '~/ui/components/BoxLink';
+
+> See the [Expo Router](./index) reference for installation and configuration.
+
+## Usage
+
+```tsx
+import { Link } from 'expo-router';
+
+export default function Page() {
+  return <Link href="/about">About</Link>;
+}
+```
+
+For more information about navigating between routes, read the navigation guide:
+
+<BoxLink
+  title="Navigate between pages"
+  Icon={BookOpen02Icon}
+  href="/router/basics/navigation/"
+  description="Learn how to navigate between pages in Expo Router."
+/>
+
+## API
+
+```js
+import { Link, Redirect } from 'expo-router';
+```
+
+<APISection packageName="expo-router/link" apiName="Link" />
