+++ docs/pages/versions/unversioned/sdk/router/native-tabs.mdx
@@ -1,5 +1,5 @@
 ---
-title: Router native tabs
+title: Native tabs
 description: An Expo Router submodule that provides native tabs layout.
 sourceCodeUrl: 'https://github.com/expo/expo/tree/main/packages/expo-router'
 packageName: 'expo-router'
@@ -16,7 +16,7 @@ import { ConfigPluginExample } from '~/ui/components/ConfigSection';
 
 `expo-router/unstable-native-tabs` is a submodule of `expo-router` and exports components to build tab layouts using platform-native system tabs.
 
-> See the [Expo Router](./router) reference for more information about the file-based routing library for native and web app.
+> See the [Expo Router](./index) reference for more information about the file-based routing library for native and web app.
 
 ## Installation
 
@@ -62,4 +62,4 @@ To learn how to use native tabs, with Expo Router, read the native tabs guide:
 import { NativeTabs } from 'expo-router/unstable-native-tabs';
 ```
 
-<APISection packageName="expo-router-native-tabs" apiName="Native tabs" />
+<APISection packageName="expo-router/native-tabs" apiName="Native tabs" />
