+++ docs/pages/versions/unversioned/sdk/router/split-view.mdx
@@ -1,5 +1,5 @@
 ---
-title: Router Split View
+title: Split View
 description: An Expo Router submodule that provides native split view layout.
 sourceCodeUrl: 'https://github.com/expo/expo/tree/main/packages/expo-router'
 packageName: 'expo-router'
@@ -20,7 +20,7 @@ import { FileTree } from '~/ui/components/FileTree';
 
 `expo-router/unstable-split-view` is a submodule of `expo-router` and exports components to build split view layouts using [platform-native system split views](https://developer.apple.com/design/human-interface-guidelines/split-views).
 
-> See the [Expo Router](./router) reference for more information about the file-based routing library for native and web app.
+> See the [Expo Router](./index) reference for more information about the file-based routing library for native and web app.
 
 ## Platform support
 
@@ -422,4 +422,4 @@ export default function Id() {
 import { SplitView } from 'expo-router/unstable-split-view';
 ```
 
-<APISection packageName="expo-router-split-view" apiName="Split view" />
+<APISection packageName="expo-router/split-view" apiName="Split view" />
