+++ docs/pages/versions/unversioned/sdk/router/ui.mdx
@@ -1,5 +1,5 @@
 ---
-title: Router UI
+title: UI
 description: An Expo Router submodule that provides headless tab components to create custom tab layouts.
 sourceCodeUrl: 'https://github.com/expo/expo/tree/main/packages/expo-router'
 packageName: 'expo-router'
@@ -14,7 +14,7 @@ import { ConfigPluginExample } from '~/ui/components/ConfigSection';
 
 `expo-router/ui` is a submodule of `expo-router` library and exports components and hooks to build custom tab layouts, rather than using the default [React Navigation](https://reactnavigation.org/) navigators provided by `expo-router`.
 
-> See the [Expo Router](./router) reference for more information about the file-based routing library for native and web app.
+> See the [Expo Router](./index) reference for more information about the file-based routing library for native and web app.
 
 ## Installation
 
@@ -55,4 +55,4 @@ For information about using `expo-router/ui` in Custom tab layouts guide:
 import { Tabs, TabList, TabTrigger, TabSlot } from 'expo-router/ui';
 ```
 
-<APISection packageName="expo-router-ui" apiName="Tabs" />
+<APISection packageName="expo-router/ui" apiName="Tabs" />
