+++ docs/public/static/data/unversioned/expo-router/split-view.json
@@ -1,0 +1,1 @@
+{"schemaVersion":"2.0","name":"expo-router/split-view","variant":"project","kind":1,"children":[{"name":"SplitViewColumnProps","variant":"declaration","kind":256,"children":[{"name":"children","variant":"declaration","kind":1024,"flags":{"isOptional":true},"type":{"type":"reference","target":{"packageName":"@types/react","packagePath":"index.d.ts","qualifiedName":"React.ReactNode"},"name":"ReactNode","package":"@types/react","qualifiedName":"React.ReactNode"}}]},{"name":"SplitViewProps","variant":"declaration","kind":256,"comment":{"summary":[{"kind":"text","text":"For full list of supported props, see ["},{"kind":"code","text":"`SplitHostProps`"},{"kind":"text","text":"](http://github.com/software-mansion/react-native-screens/blob/main/src/components/gamma/split/SplitHost.types.ts#L117)"}]},"children":[{"name":"children","variant":"declaration","kind":1024,"flags":{"isOptional":true},"type":{"type":"reference","target":{"packageName":"@types/react","packagePath":"index.d.ts","qualifiedName":"React.ReactNode"},"name":"ReactNode","package":"@types/react","qualifiedName":"React.ReactNode"}}],"extendedTypes":[{"type":"reference","target":{"packageName":"typescript","packagePath":"lib/lib.es5.d.ts","qualifiedName":"Omit"},"typeArguments":[{"type":"reference","target":{"packageName":"react-native-screens","packagePath":"src/components/gamma/split/SplitHost.types.ts","qualifiedName":"SplitHostProps"},"name":"SplitHostProps","package":"react-native-screens"},{"type":"literal","value":"children"}],"name":"Omit","package":"typescript"}]},{"name":"SplitView","variant":"declaration","kind":32,"flags":{"isConst":true},"type":{"type":"intersection","types":[{"type":"reflection","declaration":{"name":"__type","variant":"declaration","kind":65536,"signatures":[{"name":"__type","variant":"signature","kind":4096,"parameters":[{"name":"__namedParameters","variant":"param","kind":32768,"type":{"type":"reference","name":"SplitViewProps","package":"expo-router"}}],"type":{"type":"reference","target":{"packageName":"@types/react","packagePath":"index.d.ts","qualifiedName":"React.JSX.Element"},"name":"Element","package":"@types/react","qualifiedName":"React.JSX.Element"}}]}},{"type":"reflection","declaration":{"name":"__type","variant":"declaration","kind":65536,"children":[{"name":"Column","variant":"declaration","kind":1024,"type":{"type":"reflection","declaration":{"name":"__type","variant":"declaration","kind":65536,"signatures":[{"name":"__type","variant":"signature","kind":4096,"parameters":[{"name":"props","variant":"param","kind":32768,"type":{"type":"reference","name":"SplitViewColumnProps","package":"expo-router"}}],"type":{"type":"reference","target":{"packageName":"@types/react","packagePath":"index.d.ts","qualifiedName":"React.JSX.Element"},"name":"Element","package":"@types/react","qualifiedName":"React.JSX.Element"}}]}},"defaultValue":"SplitViewColumn"},{"name":"Inspector","variant":"declaration","kind":1024,"type":{"type":"reflection","declaration":{"name":"__type","variant":"declaration","kind":65536,"signatures":[{"name":"__type","variant":"signature","kind":4096,"comment":{"summary":[],"blockTags":[{"tag":"@platform","content":[{"kind":"text","text":"iOS 26+"}]}]},"parameters":[{"name":"props","variant":"param","kind":32768,"type":{"type":"reference","name":"SplitViewColumnProps","package":"expo-router"}}],"type":{"type":"reference","target":{"packageName":"@types/react","packagePath":"index.d.ts","qualifiedName":"React.JSX.Element"},"name":"Element","package":"@types/react","qualifiedName":"React.JSX.Element"}}]}},"defaultValue":"SplitViewInspector"}]}}]},"defaultValue":"..."},{"name":"SplitViewColumn","variant":"declaration","kind":64,"signatures":[{"name":"SplitViewColumn","variant":"signature","kind":4096,"parameters":[{"name":"props","variant":"param","kind":32768,"type":{"type":"reference","name":"SplitViewColumnProps","package":"expo-router"}}],"type":{"type":"reference","target":{"packageName":"@types/react","packagePath":"index.d.ts","qualifiedName":"React.JSX.Element"},"name":"Element","package":"@types/react","qualifiedName":"React.JSX.Element"}}]},{"name":"SplitViewInspector","variant":"declaration","kind":64,"signatures":[{"name":"SplitViewInspector","variant":"signature","kind":4096,"comment":{"summary":[],"blockTags":[{"tag":"@platform","content":[{"kind":"text","text":"iOS 26+"}]}]},"parameters":[{"name":"props","variant":"param","kind":32768,"type":{"type":"reference","name":"SplitViewColumnProps","package":"expo-router"}}],"type":{"type":"reference","target":{"packageName":"@types/react","packagePath":"index.d.ts","qualifiedName":"React.JSX.Element"},"name":"Element","package":"@types/react","qualifiedName":"React.JSX.Element"}}]}],"packageName":"expo-router"}
