+++ packages/expo-router/src/exports.ts
@@ -16,19 +16,6 @@ export {
 
 export { router, Router } from './imperative-api';
 
-export * from './link/Link';
-export type {
-  LinkMenuActionProps,
-  LinkMenuProps,
-  LinkPreviewProps,
-  LinkTriggerProps,
-} from './link/elements';
-export type { LinkAppleZoomProps } from './link/zoom/link-apple-zoom';
-export { usePreventZoomTransitionDismissal } from './link/zoom/usePreventZoomTransitionDismissal';
-export { type UsePreventZoomTransitionDismissalOptions } from './link/zoom/usePreventZoomTransitionDismissal.types';
-export type { DismissalBoundsRect } from './link/zoom/zoom-transition-context';
-export { useIsPreview } from './link/preview/PreviewRouteContext';
-
 export { withLayoutContext } from './layouts/withLayoutContext';
 export { Navigator, Slot };
 
@@ -58,8 +45,6 @@ export type { SingularOptions } from './useScreens';
 
 export type * from './types';
 
-export * from './color';
-
 export {
   Badge,
   BadgeProps,
@@ -71,26 +56,6 @@ export {
   VectorIconProps,
 } from './primitives';
 
-export type {
-  StackHeaderProps,
-  StackHeaderItemSharedProps,
-  StackScreenProps,
-  StackScreenBackButtonProps,
-  StackScreenTitleProps,
-  StackSearchBarProps,
-  StackToolbarBadgeProps,
-  StackToolbarButtonProps,
-  StackToolbarIconProps,
-  StackToolbarLabelProps,
-  StackToolbarMenuActionProps,
-  StackToolbarMenuProps,
-  StackToolbarProps,
-  StackToolbarSearchBarSlotProps,
-  StackToolbarSpacerProps,
-  StackToolbarViewProps,
-} from './layouts/stack-utils';
-
 export { unstable_navigationEvents } from './navigationEvents';
 
-export { Stack } from './layouts/Stack';
 export { Tabs } from './layouts/Tabs';
