+++ packages/expo-router/src/index.tsx
@@ -1,2 +1,5 @@
 import './global';
 export * from './exports';
+export * from './stack';
+export * from './link';
+export * from './color';
