+++ packages/expo-router/src/link/index.ts
@@ -1,0 +1,12 @@
+export * from './Link';
+export type {
+  LinkMenuActionProps,
+  LinkMenuProps,
+  LinkPreviewProps,
+  LinkTriggerProps,
+} from './elements';
+export type { LinkAppleZoomProps } from './zoom/link-apple-zoom';
+export { usePreventZoomTransitionDismissal } from './zoom/usePreventZoomTransitionDismissal';
+export { type UsePreventZoomTransitionDismissalOptions } from './zoom/usePreventZoomTransitionDismissal.types';
+export type { DismissalBoundsRect } from './zoom/zoom-transition-context';
+export { useIsPreview } from './preview/PreviewRouteContext';
