+++ packages/expo-router/src/native-tabs/index.ts
@@ -5,7 +5,6 @@ export type {
   XcassetIcon,
   DrawableIcon,
   MaterialIcon,
-  BaseNativeTabsTriggerIconProps,
   NativeTabsTriggerIconProps,
   NativeTabsTriggerBadgeProps,
   NativeTabsBottomAccessoryProps,
