+++ packages/expo-router/src/stack/index.ts
@@ -1,0 +1,20 @@
+export { Stack } from '../layouts/Stack';
+
+export type {
+  StackHeaderProps,
+  StackHeaderItemSharedProps,
+  StackScreenProps,
+  StackScreenBackButtonProps,
+  StackScreenTitleProps,
+  StackSearchBarProps,
+  StackToolbarBadgeProps,
+  StackToolbarButtonProps,
+  StackToolbarIconProps,
+  StackToolbarLabelProps,
+  StackToolbarMenuActionProps,
+  StackToolbarMenuProps,
+  StackToolbarProps,
+  StackToolbarSearchBarSlotProps,
+  StackToolbarSpacerProps,
+  StackToolbarViewProps,
+} from '../layouts/stack-utils';
