+++ tools/src/commands/GenerateDocsAPIData.ts
@@ -186,9 +186,12 @@ const PACKAGES_MAPPING: Record<string, CommandAdditionalParams> = {
   'expo-pedometer': ['Pedometer.ts', 'expo-sensors'],
   'expo-print': ['Print.ts'],
   'expo-router': ['exports.ts'],
-  'expo-router-ui': ['ui/index.ts', 'expo-router'],
-  'expo-router-native-tabs': ['native-tabs/index.ts', 'expo-router'],
-  'expo-router-split-view': ['split-view/index.ts', 'expo-router'],
+  'expo-router/stack': ['stack/index.ts', 'expo-router'],
+  'expo-router/link': ['link/index.ts', 'expo-router'],
+  'expo-router/color': ['color/index.ts', 'expo-router'],
+  'expo-router/native-tabs': ['native-tabs/index.ts', 'expo-router'],
+  'expo-router/split-view': ['split-view/index.ts', 'expo-router'],
+  'expo-router/ui': ['ui/index.ts', 'expo-router'],
   'expo-screen-capture': ['ScreenCapture.ts'],
   'expo-screen-orientation': ['ScreenOrientation.ts'],
   'expo-secure-store': ['SecureStore.ts'],
